[BUG FIX] Calling infer_pgsql_regular_message() in infer_pgsql_message()
Summary:
This was accidentally introduced in a previous refactoring.
Unfortunately, the existing test did not catch that. The added comment
explains it.

Also fixed a bug in infer_pgsql_regular_message(), where the last
character should be from the length field instead of the input count.

Also do not reject query without ending ';', as python API accept queries
without that.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4575

GitOrigin-RevId: 9dff9824ea74e0a9bef5fb2b203034a017290364