Minitar
Summary:
Plan is to leave headers compressed in the image and only decompress the one we want as required.

This requires `tar zxvf`. Including libarchive and minitar.

The idea would be to build minitar (460K), and include it in the container image.

Then stirling can decompress images as needed.

Test Plan: TBD

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-1863

Differential Revision: https://phab.corp.pixielabs.ai/D4626

GitOrigin-RevId: 379fa11a5da5aeac0d25b45ebb7fa321698a3052