Use compressed header packages
Summary:
We now include compressed Linux header packages, and decompress the closest match if we can't locate host headers.

Before this work started, we had included a single linux headers package consuming aboud 230 MB.

Now, we include 12 sets of linux header packages, but compressed. In total, they consume ~96 MB.

So we've saved space, and we've made things more robust.

Only downside is some initialization time is used to decompress the files, but this is minor. And it is only incurred when host header packages are not located.

Test Plan: Updated some tests, and tested manually.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1863

Differential Revision: https://phab.corp.pixielabs.ai/D4635

GitOrigin-RevId: 9fae1e5a80a4772b354ac37a25902c31a6dd271d