PC-210: fix invalid args from showing up in Live View header
Summary:
This is a surgical fix to address the issue where a url like this
https://staging.withpixie.dev:8080/live?script=px%2Fnode_stats&not_an_actual_arg=foo

will result in the header containing not_an_actual_args = foo.

It does this by setting args in context.tsx, in a filtered way where we check to make sure
that arg is either the param 'script' or an argument enumerated in vis.variables.

As stated above, this fix is surgical and I think we will need to change the data model a bit here.
It probably needs more discussion but I am imagining something like the following:
Separate the current concept of 'args' into separate concepts.
1. url params, which are the raw query params plus other non-query params such as "script".
2. query args, which are the values of url params (and later editable in the UI itself) that are valid for a given vis, excluding things like "script".
I think you could have VisWithArgs or something like that, which would contain both vis and this concept of
query args. It would also have a setArgs and setVis function on it
that handle filtering the query args to only those that are valid for the current vis.
This would fix the problem where we update vis and args in the same time in places like setScript, thus creating
a weird interaction between vis and args which are in reality very.

Test Plan: Tried various urls, we also should probably add tests for context.tsx.

Reviewers: malthus, michelle, zasgar, #engineering

Reviewed By: malthus, #engineering

JIRA Issues: PC-210

Differential Revision: https://phab.corp.pixielabs.ai/D4659

GitOrigin-RevId: 1140e149ecb6b478672a68da97a8a1191c0363db