[MySQL] Fix StmtPrepareOKResponse
Summary:
This diff fixes HandleStmtPrepareOKResponse.
The NeedsMoreData test is changed because popping off just one packet would pop off the EOF at the end. But an EOF packet is not always present. So the missing of the last EOF shouldn't trigger NeedsMoreData, but rather just a warning if not DeprecateEOF. But popping off two packets should definitely trigger NeedsMoreData.

Test Plan: To be tested with mysql_container_bpf_test.

Reviewers: oazizi, yzhao, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-1873

Differential Revision: https://phab.corp.pixielabs.ai/D4636

GitOrigin-RevId: 91ce54aae38568f09909b8dd7cc94bb8b7636553