[Cleanup] ProcTracker
Summary:
Was going through some code, and observed that we don't really need to have `absl::flat_hash_map<md::UPID, std::filesystem::path>` for the list of UPIDs; `absl::flat_hash_set<md::UPID>` suffices since the mapping from upid to path is deterministic, and can be generated on the fly.

This avoids storing the path, and copying it every time the map was previously copied.

Also, in some cases (e.g. jvm_stats_connector), the path is not used at all, so in that case it was just eating up space.

Test Plan: Existing tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4672

GitOrigin-RevId: bf3f59d7026d4f10bc4fcaf55851a73a2b415edf