Add ebnf parser for formatted string in CLI
Summary:
When the CLI/UI make an autocomplete request, they receive a formatted string of a form similar to `${1:run} ${2:script:px/svc_info} ${3:pod:pl/frontend$0}`. they should be able to parse this to determine the tabstops, labels, and values.
we use an ebnf parser to do so in the CLI. In order to making the parsing much, much simpler, i turned tabstops without a label/value from the format `$3` to `${3}`.
the parser parses each tabstop into an index number, label, and value.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4663

GitOrigin-RevId: bd40c5c3ec28655ae21b67ef7328c1bf6726dc92