[MySQL] ColDefinition parsing
Summary: This diff adds support for processing column definition packets. They are useful as part of the response of StmtExecute. In the following diff, I will parse out the type of the fields from the column definitions before the BinaryResultsetRows, and use the types to parse out the ResultsetRow values, which would prevent confusion with OK packets observed before.

Test Plan: Added Test for ProcessColDefinition

Reviewers: oazizi, yzhao, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-1890

Differential Revision: https://phab.corp.pixielabs.ai/D4684

GitOrigin-RevId: d7f4aad9b6ac08f1d187c0af0eeaf06e1a43e679