Filter out malformed /proc/pid/exe symlinks
Summary:
This causes a warning: Cannot analyze binary for uprobe deployment.

Tried for a couple hours last night to get to the root cause, but the PIDs that cause this warning appear to be very short lived.
I suspect a race, perhaps when the pid is being created or destroyed.

Test Plan: Existing tests.

Reviewers: yzhao, #engineering, zasgar

Reviewed By: #engineering, zasgar

JIRA Issues: PP-1888

Differential Revision: https://phab.corp.pixielabs.ai/D4687

GitOrigin-RevId: f001fe99cd2d2cda9db2d25d2dc90c03f68cdb3b