Fixes for autocomplete component.
Summary:
The error we keep seeing in sentry turns out to be a bug in the autocomplete component, now it wraps around correctly on keyboard up key.

Also make the limit the height of the autocomplete box with vh and move the box up a little higher now that we have more suggestions by default.

Test Plan: tested the key up wrap around no longer errors.

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-222

Differential Revision: https://phab.corp.pixielabs.ai/D4700

GitOrigin-RevId: a6026ed42c6be3bbef9fb516abd8019575146956