PP-1809: Metadata service should cache and return agent k8s entity bloom filters
Summary:
Depends on D4477. We are constructing bloom filters that contain all of the k8s entities
that a given agent has stored on it. Those bloom filters will be created by the agents, sent to the
metadata service with the agent heartbeats, and queried by the query broker when a query is executed.
This diff creates a place for the wrapper of those bloom filters, AgentDataInfo, in the metadata store.
AgentDataInfo will eventually store more than just information about K8s metadata, it will also store
things like which columns in an agent table contain which categorical values, etc.

Test Plan: added

Reviewers: michelle, zasgar, philkuz, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-1809

Differential Revision: https://phab.corp.pixielabs.ai/D4651

GitOrigin-RevId: a49b14095d97f76b1fb0a5941132b44f611302b0