Update the URL with script args.
Summary:
This populates the script args in the URL. If args were not present in the url originally, use the default values from the vis spec.

NOTE: The URL link is not fully sharable yet, because it does not contain the script ID so the recipient of the URL might not have the right script. And we can't add the script ID in the URL right now because it would wipe out any local edits to the script.

Test Plan: tested manually with a bunch of different scripts.

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-216

Differential Revision: https://phab.corp.pixielabs.ai/D4704

GitOrigin-RevId: 033cf2959aec992f0539a1df48b89f9daffa2bd7