PC-213: Automatically start up locust load testing for px-sock-shop.
Summary:
TSIA. Please note that this takes away the locust web ui because batch mode and web mode are not mutually compatible.
However, it seems like a better experience for users to not have to do anything, even if it means they lose a bit of configurability
about the load amount.

Test Plan: ran it

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-213

Differential Revision: https://phab.corp.pixielabs.ai/D4703

GitOrigin-RevId: 991bdb77717481eb5f41d023e41de282279252fc