Autocomplete: implicit "run" action shouldnt be prepended to output string
Summary:
when a user is entering a command, if they don't type "run" or "go", we assume that they implicitly mean to "run" a script.
if the action is implicit, we don't need to add it to the beginning of the string.
With the current model, this is what happens:
User types: "a"
We return: "run a"
Instead, we should just return "a", because we already assume they mean "run"

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4688

GitOrigin-RevId: 6703931e06242ed49bab19b509308ddd496bc0b2