Autocomplete: empty input should still show suggestions
Summary:
when the input is empty, we are sending a request to elastic to query for all entities with a name matching the empty string, which matches with nothing.
instead, if the input is empty, we should not apply the constraint to match on name.

Test Plan: update unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4711

GitOrigin-RevId: 06651eda328dcb18a22b7d35ec1afd7456d509f5