Autocomplete: fix ebnf parsing for tabstops with label and no value
Summary:
when playing around with autocomplete in the CLI, i found a couple bugs with the ebnf parsing for handling cases where a tabstop has a label, but no value. for example: "script:px/service_stats svc_name:" would not be parsed properly and result in an error.
we also need to allow "-" in the values, so added "-" as an Ident

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4712

GitOrigin-RevId: ffe78693a5d9c70a3563efed848e0e184c1a7d53