Autocomplete: create empty tabstop after user selects suggestion
Summary:
if a script is defined in the cmd, and there are invalid args that still need to be filled out, we move the user's cursor to that tabstop.
however, if there is no script specified yet, if a user selects a suggestion, we should move their cursor to a new, empty tabstop.

Test Plan: unit test + try in CLI

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4713

GitOrigin-RevId: c3fe01058147e6793d1eeb0ae2f1058b41bb85dd