Autocomplete: don't allow invalid labels
Summary:
if a user manually enters an invalid label (like "test:"), we were able to detect that and add it to the value of the tabstop instead of the label.
it was possible to have fields that looked like: "svc:test:abcd". however, that actually makes the ebnf parsing more difficult since it considers the ":" as a special character.
we should just not allow users to include invalid labels in the value.

if this ends up being behavior that we find strange, we can always consider changing it in the future.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4720

GitOrigin-RevId: 258fee6a43d44d1f3115582b193ebc6477d45e46