Autocomplete: add scripts to suggester results
Summary:
we need to return scripts in the suggestion results, otherwise autocomplete is unusable.
unfortunately, we won't be able to query elastic for scripts until the script indexer is done.
this solution is temporary, and uses the bundle reader to read the scripts from the bundle and parse the args. matching scripts are appended to the results, but are not ordered with respect to the non-script results.

Test Plan: ran cli autocomplete

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4725

GitOrigin-RevId: db4c93795835e761bf3324b01d005c23c5091984