Autocomplete: add input field widget
Summary:
we need an input field widget that handles scrolling/keyboard shortcuts.
tview does have one, but it is limited because it:
- doesn't expose cursor position
- doesn't allow text color changes aside from a single label/placeholder

this component is mostly copied from their implementation, since the logic to handle text scrolling is relatively complex. it has been slightly modified to allow for our usecase.

Test Plan: used in autocomplete CLI

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4730

GitOrigin-RevId: c6ab54385bac99b5c48215bc6e82948fd85128bf