Add new CLI autocomplete
Summary:
this diff adds the new cli autocomplete to the CLI live view. you can specify -n as a flag for the new mode. otherwise, the old mode is used as default.
this will only work if there are some items in the elastic index.
there are lots of improvements/tweaks that can still be made, but this is just an initial check-in so we can start playing around with it.
some things that still need to be done or thought about:
- highlighting of matched indices,
- displaying default values

Test Plan: played around with it a bunch of times

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4732

GitOrigin-RevId: ca09e953167376848b484999a9e02f6740ae3e5d