PP-1809: Add new endpoint to Vizier metadata service to fetch agent table metadata
Summary:
Query broker needs to know the schema that each agent has today, so when it runs a query it calls GetSchemas.
We are adding information about the k8s entities that each agent has to the compiler, so that the compiler can do smart
pruning of the query. That information is stored in a struct called AgentDataInfo, which can be extended in the future
to inform the compiler about any other properties of the data that a particular agent has.
THe plan for this new endpoint is to subsume GetSchemas from the query broker's perspective, and get both schemas
and the data info for each agent together in one call.

Depends on D4651.

Test Plan: added tests

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-1809

Differential Revision: https://phab.corp.pixielabs.ai/D4655

GitOrigin-RevId: abfedfb0dc8e937129f4ecd9bace7908645338e4