Fix Legend Data float64 render
Summary: NumeralJS is so aggravating. This is the second time numeral has messed with us. Re-using the formatter we use for tables to avoid this problem.

Test Plan: Tested in dev with JVM stats which currently has no full gc events (unrelated, might be a bug) but yields 0 values that look like ridiculous floating point values. Numeral JS decides that's a NaN.

Reviewers: jamesbartlett, malthus, #engineering

Reviewed By: malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4741

GitOrigin-RevId: 8822cd1a2837b90f6066d94fc2547a36dd2a8a66