PP-1811: Initial implementation of filter pushdown
Summary:
We want to move filters as early in the query as possible for effiency reasons.
This diff implements a rule to do that, which only works on certain kinds of operators for now.
This diff has a small refactor a utility in ir_nodes.h so that it could be used in rules.cc as well.
Next up will be supporting pushing filters past aggs, joins, and unions.

Test Plan: added

Reviewers: philkuz, jamesbartlett, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-1811

Differential Revision: https://phab.corp.pixielabs.ai/D4665

GitOrigin-RevId: 4e829528592397c3ea83ff998d32d20c00869fbb