PP-1810 part 1: Lay groundwork for distributed coordinator to allocate queries only to agents that may produce data
Summary:
This diff does not actually change any existing functionality, it just refactors the current distributed coordinator
as well as the IR::Clone() method in order to support PP-1810.

In this change, the distributed coordinator looks at each subgraph of a query and evaluates whether or not a given carnot instance
will produce data for that subgraph. (In this diff, the answer is always yes, and I will add conditional filtering based on
metadata stored in the carnot instance in the next diff). If it does produce data, it is added to the plan that is executed on that
Carnot.

These functions are used only by the current coordinator which assumes that only agents produce data, but I tried to write them in a way that will be usable by other coordinator types in the future.

Test Plan: added a test for the new method to IR, will add other tests when the coordinator actually does metadata-based filtering.

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-1810

Differential Revision: https://phab.corp.pixielabs.ai/D4691

GitOrigin-RevId: 5feb653427889a3f18ca42dc2e073058ea127760