PP-1814 part 2: Metadata service should store agent data info from heartbeat in metadata store
Summary: Depends on D4744. This is the final piece on the agent side, completing the flow where agents create and send metadata filters to the metadata service, where those filters are stored in the metadata store. The compiler is then passed those filters to use for query pruning.

Test Plan: augmented existing test

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-1814

Differential Revision: https://phab.corp.pixielabs.ai/D4745

GitOrigin-RevId: 386ea527cc7290752baf1243754d3a68248c326e