When detaching, BCCWrapper ignores uprobe whose binary no longer exists
Summary:
Kernel will remove any lingering uprobe instance.
We are still not sure if BCC itself keeps any data.

But nonetheless, detaching in the situation where binary no longer exists is not necessary.
And it can happen because of container terminates, which is a legitimate scenario.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-1861

Differential Revision: https://phab.corp.pixielabs.ai/D4740

GitOrigin-RevId: ecb9793983d03da7434062b668d3480496a38ac7