Enum utils: create an enum spec
Summary:
This little utility will be useful for Stirling to encode its enum values and push them out as part of its publish proto.

Eventually, the UI can use this information to represent enum values as more human readable values.

For example, MySQL has a command field, which is currently a number that is quite useless to users. Eventually, this could be fixed.

This is just the first piece of this plan.

Test Plan: Added a test for the utility.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1898

Differential Revision: https://phab.corp.pixielabs.ai/D4701

GitOrigin-RevId: fbb4ca166b8c9260230fe7b1ce43cad8173d0cec