Add utils to convert k8s node object to our proto
Summary:
when we receive a k8s node update from the k8s api, it comes in the form of a v1.Node object.
we need to convert this object into our node proto message, for storage in etcd. this diff adds the util for doing that conversion.

Test Plan: unit test

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4751

GitOrigin-RevId: 7458b9abdd691402ee07bbe8afc870289ad03af7