Fix bug with error channel write hanging
Summary:
We used to send errors on errCh. A write to that channel blocks when there are no readers. The reader dies ~30 mins due to TCP timeout (expected).
Sometimes a write comes in at the same time and that causes and error to be written to errCh. The fact that there is an error pending prevents the
whole state from shutting down cleanly. Which is normally ok. However, it also subscribes to NATS using a queue group and we hang
onto stale subscriptions which means some of them go into a blackhole instead of going over a new connection.

Test Plan: Existing tests. Hard to test the race condition. Will check and see if it occurs on prod again.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4757

GitOrigin-RevId: 013776021210d047b9ce76e23585b4807bdd7c06