Listen and store K8s node updates in metadata
Summary:
we should be listening to k8s and storing the node updates in the metadata service.
this part just looks similar to how we store and update our other resources.
there will be a followup diff for pulling out and aggregating the pod CIDRs

Test Plan: ran skaffold, unit test

Reviewers: zasgar, yzhao, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4760

GitOrigin-RevId: 731a8fc901d1d1dd394c30257d30d89909c9ab8d