Add node update proto message
Summary:
whenever the metadata service sends an update to the agent, it sends it in the form of a ResourceUpdate. this diff adds the NodeUpdate message, which is contained within a ResourceUpdate message. essentially, this is the information that agent needs from the metadata service, and should only contain a subset of the information that's actually stored in etcd.
for the CIDR bug, the relevant field is PodCIDR/PodCIDRs

Test Plan: n/a

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4753

GitOrigin-RevId: 441268fa15884d9edc7858c63848e5f1ad677d9d