Add endpoint for autocompleting single field
Summary:
we need to add an endpoint for autocompleting single field, which doesn't deal with all of the fancy string formatting/tabstops that the normal autocomplete endpoint has.
when making a request, specify the input string that should be matched (can be empty string), the entity type that we should be matching with (svc, pod, etc). If autocompleting a script, you can also specify the required arg types, which specify the argument types that the matched scripts must take.

Test Plan: unit test

Reviewers: malthus, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4774

GitOrigin-RevId: f3686a54ed6e6415ecf240610bbe1a5d92bc4a14