PP-1808: Add k8s entities to metadata filter with namespace prepended
Summary:
In the query language, we expect pod/service names to have the namespace prepended.
Although that is not the long term solution, in this diff they get added to the metadata filter
with the namespace prepended as well as a stopgap solution until the query language is shifted
to only use the pod name itself.

Test Plan: edited

Reviewers: zasgar, michelle, philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-1808

Differential Revision: https://phab.corp.pixielabs.ai/D4785

GitOrigin-RevId: 24d89cf755b47ef4681c5227a5130c0884733f32