DwarfReader: Basic foundation
Summary:
This diff adds a basic DwarfReader library, to extract dwarf information from an object file. The basic entity that is returned is a DIE object, which has information we're looking for. The information from the DIE is not yet extracted.

A tool is also included for testing purposes. The tool primitively replicates llvm-dwarfdump.

There is a basic test, but this is just a foundation.

Test Plan: Basic test case added.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1367

Differential Revision: https://phab.corp.pixielabs.ai/D4765

GitOrigin-RevId: 77638d8b8a2e72a5483cafa4cb6c2fd6a5866682