Autocomplete: add parser for autocomplete format string
Summary:
the autocomplete backend returns a formatted string which indicates the different tabstops/fields in the command.
the tab indexes indicate the order to tab through the different fields. the backend also gives a map of suggestions, where the key is the tabstop index, and the value contains the suggestions for that tabstop.
the tab label is the label for the tabstop. not all tabstops will have a label.
the tab value is the value for the tabstop, which the user can edit.

the CLI uses an ebnf parser to parse the formatted string. however, it looks like javascript doesnt have many good ebnf parsers (50-60 stars on github). the string is also simple enough to use regex, so that is what i use here.

with the tabstops in nice structs, hopefully this makes building the autocomplete in the UI somewhat less daunting.

Test Plan: unit test

Reviewers: malthus, zasgar, #engineering

Reviewed By: malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4794

GitOrigin-RevId: 75cbdbb17890c58662ca28171d65ba4d7199a5e5