Do not clear upid when aggregating client side events
Summary:
The original idea is to aggregate client traffic to the same remote endpoint.

It creates a bug, because later when the SocketTraceConnector uses the code below to filter out
upids that are not tracked by MDS:

<code>
for (const auto& [key, stats] : connection_stats_) {
  if (!upids.contains(key.upid)) {
    continue;
  }
  ...
}
</code>

The cleared key.upid wont match any upid tracked by MDS. This causes the record not output.

When looking at the data, it seems more useful to output upids, but aggregating by scripts.
That way the raw data appears more "correct". Prior to this diff, the raw data would show upid as
all "0", which can be a bit confusing.

Test Plan: Jenkins

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4807

GitOrigin-RevId: ddc725df2a18e51ef14db7444f20454962bdb6c5