PC-243: Add profile popover menu in Live View
Summary:
See attached screenshot of the design as well as the implementation. New popover for profile menu.
Docs button opens the docs in a new tab. Logout does logout.
Uses the auth0 profile picture and defaults to an initials-based avatar in the event that the profile picture is empty or broken.
I couldn't figure out how to get dividers to work, @malthus said he might be able to help out with that.

Test Plan: ran it in the ui

Reviewers: malthus, michelle, zasgar, #engineering

Reviewed By: malthus, #engineering

Subscribers: malthus

JIRA Issues: PC-243

Differential Revision: https://phab.corp.pixielabs.ai/D4804

GitOrigin-RevId: bc6d6a2307b989d7fe6854c963c8e798ddbfcb2f