PP-1918: Don't merge two filters/limits when their relations differ
Summary:
We were previously merging filters and limits when their output relations
were different. Because filters and limits strictly mirror their parent relations, this caused columns
to show up in unexpected indices in Carnot's outputs. In the cases where the output relations are not
shared, we shouldn't merge those nodes for now. In the future we can add a map node after the limit
if we want to in order to reuse the same limit/filter node to produce different relations, but for now
it doesn't seem like a huge priority.

Test Plan: added

Reviewers: philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-1918

Differential Revision: https://phab.corp.pixielabs.ai/D4834

GitOrigin-RevId: 7b9e08b8d3e6eaee57c905cde3f80dd852f0173f