Fix bug where banner doesn't show up for Pixie support account
Summary: the recent update to the userresolver introduced a bug where the pixie support account banner was not showing up. this is because the pixie support account is not a real user in pixie, so the GQL request would throw an error. instead, we shouldn't throw an error for accounts with no users, but still return non-userInfo fields such as orgname (which is used to determine whether to show the banner)

Test Plan: unit test, ran in skaffold

Reviewers: zasgar, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4840

GitOrigin-RevId: 1fc140ce4f4dd9bf32ea7b7b69b762f4723580f2