Reenable scripts for autocomplete
Summary:
in D4835. i had to temporarily comment out the bundle manager because changes broke the API server.
adding it back in here so we can continue our progress on autocomplete. hopefully soon we will have script indexing so that we dont need to use the bundle manager in the API service anymore.

Test Plan: ran skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4842

GitOrigin-RevId: 5cf74801e5ab723819b79b355be0b6aa2cdd9d9e