Verify vizier version before sending update request
Summary:
users can hit a bad state if they try to update to a version that does not exist. we blindly accept the version string and send the request to the cloud connector, which leads to an image pull backoff for the updater job.
instead, we should verify that the version string is a valid vizier version.
this does not prevent all possible update failure cases, but does prevent one that we keep running into during testing.

Test Plan: ran in skaffold and tried updating to both valid and invalid versions

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4847

GitOrigin-RevId: 1c51fe1cbe4aa734dfbb9860641850c24d4853fd