[Cleanup] SocketTraceBPFTest parameter for client-side tracing
Summary:
Whether the client-side tracing kicks in or not depends on the CIDR, but that's a detail the tests should not have to worry about.

So the SocketTraceBPFTest now takes a template parameter to specify whether client-side tracing is enabled by the tracer.

Note that server-side tracing is always active.

Also threw in some coverage to the two new columns we've added.

Test Plan: This diff is about tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4838

GitOrigin-RevId: 15ec4d69f43d6b919bece99e739f29d5bcda5d7a