DwarfReader: simplifications
Summary:
Cleaning up the code a bit, and only consider short name, not linkage name.

The linkage name appears to be the mangled name, but we probably don't need that for the work we're doing.

So removing it. We'll see what happens. If we need it back, easy to add back in.

Test Plan: Existing dwarf tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4875

GitOrigin-RevId: 898ba0f1bcab123ecf442edf0762d62e9291ef65