[PP-1917] Drop Columns in the Limit node
Summary: In this diff we allow the Limit node to drop columns. This change comes about because we have support for dropping columns in Filter already and there are cases where we don't really need another op in front of Limit. Alternatively, we could have placed a Map in front of limit and in fact I have a way to do this. However we would save on evaluation of expressions if we just drop the columns here. The limit operator also has a specification field for columns so this should work anyways.

Test Plan: Tested limit changes in limit_node_test and made sure that this works in skaffold

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4876

GitOrigin-RevId: 369408454206c52d65c72615a6d4127c57383381