PP-1928, PP-1916: Update analyzer to use new metadata handling
Summary: Depends on D4877. This diff completes the refactor of how metadata is handled in the compiler. It is converted to a metadata generating function which has annotations attached to it labeling what kind of metadata type it is. These annotations are passed down when the output of that func is assigned to another column name or used in a group by clause or filter. These annotations are the bases for agent metadata filtering. This diff removes the old behavior where metadata columns were created via a map and called _attr_<metadata_name>. Next diff I will delete the obsolete classes.

Test Plan: added/existing

Reviewers: philkuz, jamesbartlett, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-1928, PP-1916

Differential Revision: https://phab.corp.pixielabs.ai/D4882

GitOrigin-RevId: 0ea09742110cbd31e5ee86785584f67535eccc6e