Only parse protobuf if the content type includes 'application/grpc'.
Summary:
Why originally we thought it wont be problematic to always parse http2 payload as protobuf?
It seems just mental bounded to gRPC during the development.

Obviously, not all HTTP2 are for gRPC.

This removes some cases of misclassification, like reported in:
https://pixie-labs.slack.com/archives/CC34SPKJM/p1589940966033800?thread_ts=1589931211.028200&cid=CC34SPKJM

But for https://pixie-labs.atlassian.net/browse/PP-1897, there are other cases of failures, which
are being investigated.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Subscribers: nserrino

JIRA Issues: PP-1897

Differential Revision: https://phab.corp.pixielabs.ai/D4896

GitOrigin-RevId: edfda4c4771a28b174a0fd9cccf72e446d11a8c6