Fix for args remaining in the URL when swtiching scripts.
Summary:
The change to preserve the query params was added in https://phab.corp.pixielabs.ai/rPLM885a4614b73b5fdb89c78580cb903026c1d241a9 in order to preseve the orgname. But that also caused the query params to linger when switching between scripts. Change back to always overwriting the query params since we are not using the orgname in the URL anymore.

- move args context functions into a utils module.

Test Plan: Added tests for the utils functions

Reviewers: michelle, #engineering, nserrino

Reviewed By: #engineering, nserrino

Subscribers: nserrino

JIRA Issues: PC-269

Differential Revision: https://phab.corp.pixielabs.ai/D4894

GitOrigin-RevId: a55398385fd06a0cc81864fb5d318e8006d47df5